pyproject.toml
[project]
name = "pql-agent"
version = "0.1.0"
description = "Celonis Helper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "chromadb>=1.5.5",
  "openai>=2.31.0",
  "pre-commit>=4.5.1",
  "python-dotenv>=1.2.2",
  "requests>=2.31.0",
  "tiktoken>=0.12.0",
]

[project.scripts]
pql-agent = "pql_agent.cli:main"

[project.optional-dependencies]
notebook = [
  "ipykernel>=6.29.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.1.0",
  "seaborn>=0.13.0",
  "tiktoken>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
